backend/telemetry-analysis/requirements.txt
annotated-types==0.7.0
dependency-injector==4.41.0
dnspython==2.6.1
grpcio==1.64.1
grpcio-tools==1.64.1
protobuf==5.27.2
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
six==1.16.0
typing_extensions==4.12.2
